+++ lib/recipes/college-success/collegesuccess_recipe.rb
@@ -11,7 +11,7 @@ COLLEGESUCCESS_RECIPE = Kitchen::BookRecipe.new(
 
   book.search('cnx-pi').trash
 
-  BakeImages.v1(book: book, resources: resources)
+  # BakeImages.v1(book: book, resources: resources)
   BakeUnnumberedFigure.v1(book: book)
   BakePreface.v1(book: book)
   book.pages('$.preface').each do |page|