+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+* Bake `figures in preface` in `college-success`
+
 ## [1.32.0] - 2023-06-02
 
 * Suppress `High School Features` in `college-success`